refactor(docs): textarea dx
