+++ apps/docs/content/components/textarea/autosize.ts
@@ -1,25 +1,4 @@
-const App = `import {Textarea} from "@nextui-org/react";
-
-export default function App() {
-  return (
-    <div className="flex w-full flex-wrap md:flex-nowrap mb-6 md:mb-0 gap-4">
-      <Textarea
-        label="Description"
-        placeholder="Enter your description (Default autosize)"
-      />
-      <Textarea
-        minRows={2}
-        label="Description"
-        placeholder="Enter your description (Min rows 2)"
-      />
-      <Textarea
-        maxRows={3}
-        label="Description"
-        placeholder="Enter your description (Max rows 3)"
-      />
-    </div>
-  );
-}`;
+import App from "./autosize.raw.jsx?raw";
 
 const react = {
   "/App.jsx": App,
