+++ apps/docs/content/components/textarea/controlled.ts
@@ -1,22 +1,4 @@
-const App = `import {Textarea} from "@nextui-org/react";
-
-export default function App() {
-  const [value, setValue] = React.useState("");
-
-  return (
-    <div className="w-full flex flex-col gap-2 max-w-[240px]">
-      <Textarea
-        variant="underlined"
-        label="Description"
-        labelPlacement="outside"
-        placeholder="Enter your description"
-        value={value}
-        onValueChange={setValue}
-      />
-      <p className="text-default-500 text-small">Textarea value: {value}</p>
-    </div>
-  );
-}`;
+import App from "./controlled.raw.jsx?raw";
 
 const react = {
   "/App.jsx": App,
