+++ apps/docs/content/components/textarea/description.raw.jsx
@@ -1,0 +1,13 @@
+import {Textarea} from "@nextui-org/react";
+
+export default function App() {
+  return (
+    <Textarea
+      className="max-w-xs"
+      description="Enter a concise description of your project."
+      label="Description"
+      placeholder="Enter your description"
+      variant="faded"
+    />
+  );
+}
