+++ apps/docs/content/components/textarea/description.ts
@@ -1,16 +1,4 @@
-const App = `import {Textarea} from "@nextui-org/react";
-
-export default function App() {
-  return (
-    <Textarea
-      variant="faded"
-      label="Description"
-      placeholder="Enter your description"
-      description="Enter a concise description of your project."
-      className="max-w-xs"
-    />
-  );
-}`;
+import App from "./description.raw.jsx?raw";
 
 const react = {
   "/App.jsx": App,
