+++ apps/docs/content/components/textarea/disable-autosize.raw.jsx
@@ -1,0 +1,17 @@
+import {Textarea} from "@nextui-org/react";
+
+export default function App() {
+  return (
+    <Textarea
+      disableAnimation
+      disableAutosize
+      classNames={{
+        base: "max-w-xs",
+        input: "resize-y min-h-[40px]",
+      }}
+      label="Description"
+      placeholder="Enter your description"
+      variant="bordered"
+    />
+  );
+}
