+++ apps/docs/content/components/textarea/disable-autosize.ts
@@ -1,20 +1,4 @@
-const App = `import {Textarea} from "@nextui-org/react";
-
-export default function App() {
-  return (
-    <Textarea
-      label="Description"
-      variant="bordered"
-      placeholder="Enter your description"
-      disableAnimation
-      disableAutosize
-      classNames={{
-        base: "max-w-xs",
-        input: "resize-y min-h-[40px]",
-      }}
-    />
-  );
-}`;
+import App from "./disable-autosize.raw.jsx?raw";
 
 const react = {
   "/App.jsx": App,
