+++ apps/docs/content/components/textarea/error-message.ts
@@ -1,18 +1,4 @@
-const App = `import {Textarea} from "@nextui-org/react";
-
-export default function App() {
-  return (
-    <Textarea
-      isInvalid={true}
-      variant="bordered"
-      label="Description"
-      placeholder="Enter your description"
-      defaultValue="NextUI is a React UI library with..."
-      errorMessage="The description should be at least 255 characters long."
-      className="max-w-xs"
-    />
-  );
-}`;
+import App from "./error-message.raw.jsx?raw";
 
 const react = {
   "/App.jsx": App,
