+++ apps/docs/content/components/textarea/readonly.raw.jsx
@@ -1,0 +1,15 @@
+import {Textarea} from "@nextui-org/react";
+
+export default function App() {
+  return (
+    <Textarea
+      isReadOnly
+      className="max-w-xs"
+      defaultValue="NextUI is a React UI library that provides a set of accessible, reusable, and beautiful components."
+      label="Description"
+      labelPlacement="outside"
+      placeholder="Enter your description"
+      variant="bordered"
+    />
+  );
+}
