+++ apps/docs/content/components/textarea/readonly.ts
@@ -1,18 +1,4 @@
-const App = `import {Textarea} from "@nextui-org/react";
-
-export default function App() {
-  return (
-    <Textarea
-      isReadOnly
-      label="Description"
-      variant="bordered"
-      labelPlacement="outside"
-      placeholder="Enter your description"
-      defaultValue="NextUI is a React UI library that provides a set of accessible, reusable, and beautiful components."
-      className="max-w-xs"
-    />
-  );
-}`;
+import App from "./readonly.raw.jsx?raw";
 
 const react = {
   "/App.jsx": App,
