+++ apps/docs/content/components/textarea/required.raw.jsx
@@ -1,0 +1,13 @@
+import {Textarea} from "@nextui-org/react";
+
+export default function App() {
+  return (
+    <Textarea
+      isRequired
+      className="max-w-xs"
+      label="Description"
+      labelPlacement="outside"
+      placeholder="Enter your description"
+    />
+  );
+}
