+++ apps/docs/content/components/textarea/required.ts
@@ -1,16 +1,4 @@
-const App = `import {Textarea} from "@nextui-org/react";
-
-export default function App() {
-  return (
-    <Textarea
-      isRequired
-      label="Description"
-      labelPlacement="outside"
-      placeholder="Enter your description"
-      className="max-w-xs"
-    />
-  );
-}`;
+import App from "./required.raw.jsx?raw";
 
 const react = {
   "/App.jsx": App,
