+++ apps/docs/content/components/textarea/usage.raw.jsx
@@ -1,0 +1,5 @@
+import {Textarea} from "@nextui-org/react";
+
+export default function App() {
+  return <Textarea className="max-w-xs" label="Description" placeholder="Enter your description" />;
+}
