+++ apps/docs/content/components/textarea/usage.ts
@@ -1,14 +1,4 @@
-const App = `import {Textarea} from "@nextui-org/react";
-
-export default function App() {
-  return (
-    <Textarea
-      label="Description"
-      placeholder="Enter your description"
-      className="max-w-xs"
-    />
-  );
-}`;
+import App from "./usage.raw.jsx?raw";
 
 const react = {
   "/App.jsx": App,
